requirements/_base.txt
bpython==0.12
South==0.7.6
django-reversion=1.6
